requirements.txt
fastapi~=0.115.0
pydantic~=2.9.2
sqlalchemy~=2.0.35
uvicorn~=0.31.0
pydantic_settings~=2.5.2
alembic~=1.13.3
python-dotenv~=1.0.1
asyncpg~=0.29.0
passlib~=1.7.4
bcrypt~=4.2.0
starlette~=0.45.3
pytest~=8.3.4
httpx~=0.28.1
python-jose==3.3.0
loguru==0.7.3